fix: fixed infinite latitude and longitude
There was an issue that the files had in that there are some points in the files that are not actually on the earth due to converage and array sizes.  This can cause problems when plotting as pcolormesh doesn't accept masked or non-finite valued arrays.  I simply filled the inf values with 1e30 so that they plot.